Add FMOperator module
The FMOperator is a Quad (4 voices) VCO designed to be used for FM synthesis. It is meant to be used in the way of "operators" as popularized by Yamaha's DX7, except that it doesn't carry its own envelope generator. Instead, envelopes should be provided externally. The design encourages chaining (stacking) operators, which is how you can make interesting FM sounds.